# Add missing indexes to spree_orders and spree_products