--- a/db/migrate/20200404080853_add_user_id_index_to_spree_orders.rb
+++ b/db/migrate/20200404080853_add_user_id_index_to_spree_orders.rb
@@ -1,0 +1,5 @@
+class AddUserIdIndexToSpreeOrders < ActiveRecord::Migration
+  def change
+    add_index :spree_orders, :user_id
+  end
+end