--- a/db/migrate/20200404081018_add_supplier_id_index_to_spree_products.rb
+++ b/db/migrate/20200404081018_add_supplier_id_index_to_spree_products.rb
@@ -1,0 +1,5 @@
+class AddSupplierIdIndexToSpreeProducts < ActiveRecord::Migration
+  def change
+    add_index :spree_products, :supplier_id
+  end
+end