#### What? Why?

Closes #5137 
Closes #4424

<!-- Explain why this change is needed and the solution you propose.
Provide context for others to understand it. -->

Adds previously identified missing indexes to spree_orders and spree_products.

#### What should we test?
<!-- List which features should be tested and how. -->

Improved query plans.

#### Release notes
<!-- Write a line or two to be included in the release notes.
Everything is worth mentioning, because you did it for a reason. -->

Added database indexes on products and orders

<!-- Please assign one category to your PR and delete the others. 
The categories are based on https://keepachangelog.com/en/1.0.0/. -->

Changelog Category: Added