@Matt-Yorkley 2020-04-04
I also had a go with an automated missing indexes scan, using the [lol_dba](https://github.com/plentz/lol_dba) gem, and it recommended this (on top of the two indexes added here):

```ruby
class AddMissingIndexes < ActiveRecord::Migration
    def change
        add_index :column_preferences, :user_id
        add_index :coordinator_fees, [:enterprise_fee_id, :order_cycle_id]
        add_index :customers, :enterprise_id
        add_index :distributors_payment_methods, [:distributor_id, :payment_method_id]
        add_index :enterprise_groups_enterprises, [:enterprise_group_id, :enterprise_id]
        add_index :enterprise_groups_enterprises, [:enterprise_id, :group_id]
        add_index :exchanges, [:enterprise_id, :order_cycle_id]
        add_index :inventory_items, :enterprise_id
        add_index :inventory_items, :variant_id
        add_index :order_cycle_schedules, [:order_cycle_id, :schedule_id]
        add_index :order_cycles, :coordinator_id
        add_index :proxy_orders, :order_cycle_id
        add_index :spree_activators, [:id, :type]
        add_index :spree_addresses, :country_id
        add_index :spree_addresses, :state_id
        add_index :spree_adjustments, :originator_id
        add_index :spree_adjustments, [:adjustable_id, :adjustable_type]
        add_index :spree_adjustments, [:originator_id, :originator_type]
        add_index :spree_adjustments, [:source_id, :source_type]
        add_index :spree_assets, [:id, :type]
        add_index :spree_assets, [:viewable_id, :viewable_type]
        add_index :spree_calculators, [:calculable_id, :calculable_type]
        add_index :spree_calculators, [:id, :type]
        add_index :spree_inventory_units, :return_authorization_id
        add_index :spree_line_items, :tax_category_id
        add_index :spree_log_entries, [:source_id, :source_type]
        add_index :spree_option_values, :option_type_id
        add_index :spree_option_values_line_items, [:option_value_id, :spree/line_item_id]
        add_index :spree_option_values_variants, [:option_value_id, :spree/variant_id]
        add_index :spree_option_values_variants, [:spree/option_value_id, :variant_id]
        add_index :spree_orders, :bill_address_id
        add_index :spree_orders, :created_by_id
        add_index :spree_orders, :ship_address_id
        add_index :spree_payment_methods, :provider_id
        add_index :spree_payment_methods, [:id, :type]
        add_index :spree_payments, :payment_method_id
        add_index :spree_payments, [:source_id, :source_type]
        add_index :spree_product_option_types, :option_type_id
        add_index :spree_product_option_types, :product_id
        add_index :spree_product_properties, :property_id
        add_index :spree_products, :shipping_category_id
        add_index :spree_products, :tax_category_id
        add_index :spree_products_promotion_rules, [:product_id, :promotion_rule_id]
        add_index :spree_products_promotion_rules, [:promotion_rule_id, :spree/product_id]
        add_index :spree_promotion_action_line_items, :promotion_action_id
        add_index :spree_promotion_action_line_items, :variant_id
        add_index :spree_promotion_actions, :activator_id
        add_index :spree_promotion_actions, [:id, :type]
        add_index :spree_promotion_rules, :activator_id
        add_index :spree_promotion_rules, [:id, :type]
        add_index :spree_promotion_rules_users, [:promotion_rule_id, :user_id]
        add_index :spree_return_authorizations, :order_id
        add_index :spree_roles_users, [:spree/role_id, :user_id]
        add_index :spree_roles_users, [:spree_role_id, :user_id]
        add_index :spree_shipments, :address_id
        add_index :spree_shipments, :stock_location_id
        add_index :spree_shipping_methods_zones, [:shipping_method_id, :zone_id]
        add_index :spree_shipping_rates, :shipment_id
        add_index :spree_shipping_rates, :shipping_method_id
        add_index :spree_state_changes, :user_id
        add_index :spree_state_changes, [:stateful_id, :stateful_type]
        add_index :spree_states, :country_id
        add_index :spree_stock_locations, :country_id
        add_index :spree_stock_locations, :state_id
        add_index :spree_stock_movements, [:originator_id, :originator_type]
        add_index :spree_tax_rates, :tax_category_id
        add_index :spree_tax_rates, :zone_id
        add_index :spree_users, :bill_address_id
        add_index :spree_users, :ship_address_id
        add_index :spree_variants, [:spree/image_id, :spree/product_id]
        add_index :spree_zone_members, :zone_id
        add_index :spree_zone_members, [:zoneable_id, :zoneable_type]
        add_index :tag_rules, :enterprise_id
        add_index :tag_rules, [:id, :type]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :customer_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :exchange_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/billing_integration_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/gateway/bogu_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/gateway/bogus_simple_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/gateway/mig_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/gateway/pay_pal_express_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/gateway/pin_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/gateway/stripe_connect_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/gateway/stripe_sca_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/gateway_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/payment_method/check_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/payment_method_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :spree/shipping_method_id]
        add_index :taggings, [:acts_as_taggable_on/tag_id, :variant_override_id]
        add_index :variant_overrides, :hub_id
        add_index :variant_overrides, :variant_id
    end
end
```