@luisramos0 2020-04-04
:-O
Adding indexes is not always the best approach, right?
Indexes take space (they are basically simplified tables maintained in parallel to the actual table. They make inserts slower.
Having said this I am sure some of these would be very beneficial! How do we choose?