@mkllnk 2020-04-17
Looking good. :+1: 

I tested on UK staging because there is a lot of data. I'm not very experienced in analysing database performance. So I hope this makes sense and proofs that this PR is an improvement.

Before deploying:
```
puts Spree::Order.where(user_id: 207078).explain
  Spree::Order Load (8.6ms)  SELECT "spree_orders".* FROM "spree_orders" WHERE "spree_orders"."user_id" = 207078
  EXPLAIN (0.7ms)  EXPLAIN SELECT "spree_orders".* FROM "spree_orders" WHERE "spree_orders"."user_id" = 207078
EXPLAIN for: SELECT "spree_orders".* FROM "spree_orders"  WHERE "spree_orders"."user_id" = 207078
                                                               QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------------------
 Index Scan using spree_orders_completed_at_user_id_created_by_id_created_at_idx on spree_orders  (cost=0.42..7661.20 rows=21 width=209)
   Index Cond: (user_id = 207078)
(2 rows)


puts Spree::Product.where(supplier_id: 1).explain
  Spree::Product Load (4.8ms)  SELECT "spree_products".* FROM "spree_products" WHERE "spree_products"."supplier_id" = 1 AND ("spree_products".deleted_at IS NULL)
  EXPLAIN (0.6ms)  EXPLAIN SELECT "spree_products".* FROM "spree_products" WHERE "spree_products"."supplier_id" = 1 AND ("spree_products".deleted_at IS NULL)
EXPLAIN for: SELECT "spree_products".* FROM "spree_products"  WHERE "spree_products"."supplier_id" = 1 AND ("spree_products".deleted_at IS NULL)
                            QUERY PLAN
------------------------------------------------------------------
 Seq Scan on spree_products  (cost=0.00..762.89 rows=5 width=290)
   Filter: ((deleted_at IS NULL) AND (supplier_id = 1))
(2 rows)
```

After the deploy:
```
puts Spree::Order.where(user_id: 207078).explain
  Spree::Order Load (2.4ms)  SELECT "spree_orders".* FROM "spree_orders" WHERE "spree_orders"."user_id" = 207078
  EXPLAIN (1.1ms)  EXPLAIN SELECT "spree_orders".* FROM "spree_orders" WHERE "spree_orders"."user_id" = 207078
EXPLAIN for: SELECT "spree_orders".* FROM "spree_orders"  WHERE "spree_orders"."user_id" = 207078
                                              QUERY PLAN
------------------------------------------------------------------------------------------------------
 Index Scan using index_spree_orders_on_user_id on spree_orders  (cost=0.42..64.03 rows=21 width=209)
   Index Cond: (user_id = 207078)
(2 rows)


puts Spree::Product.where(supplier_id: 1).explain
  Spree::Product Load (1.3ms)  SELECT "spree_products".* FROM "spree_products" WHERE "spree_products"."supplier_id" = 1 AND ("spree_products".deleted_at IS NULL)
  EXPLAIN (0.6ms)  EXPLAIN SELECT "spree_products".* FROM "spree_products" WHERE "spree_products"."supplier_id" = 1 AND ("spree_products".deleted_at IS NULL)
EXPLAIN for: SELECT "spree_products".* FROM "spree_products"  WHERE "spree_products"."supplier_id" = 1 AND ("spree_products".deleted_at IS NULL)
                                            QUERY PLAN
--------------------------------------------------------------------------------------------------
 Bitmap Heap Scan on spree_products  (cost=4.34..30.10 rows=5 width=290)
   Recheck Cond: (supplier_id = 1)
   Filter: (deleted_at IS NULL)
   ->  Bitmap Index Scan on index_spree_products_on_supplier_id  (cost=0.00..4.34 rows=7 width=0)
         Index Cond: (supplier_id = 1)
(5 rows)
```

The key numbers:


Order | time | cost | index
-- | -- | -- | --
before | 8.6ms | 0.42..7661.20 | spree_orders_completed_at_user_id_created_by_id_created_at_idx
after | 2.4ms | 0.42..64.03 | index_spree_orders_on_user_id on spree_orders
  |   |   |  
Product | time | cost | index
before | 4.8ms | 0.00..762.89 | Seq Scan on spree_products
after | 1.3ms | 4.34..30.10 | index_spree_products_on_supplier_id